Change order_quantity type to double

+++ database/migrations/2023_10_25_094203_create_po_items_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->foreignIdFor(PurchaseOrder::class);
             $table->foreignIdFor(RopeSpec::class);
-            $table->integer('order_quantity')->comment('จำนวนที่สั่ง เช่น เชือก A สั่ง 1000 ม้วน');
+            $table->double('order_quantity')->comment('จำนวนที่สั่ง เช่น เชือก A สั่ง 1000 ม้วน');
             $table->double('unit_price')->nullable()->comment('price of spec per unit');
             $table->string('unit')->nullable()->comment('หน่วยของเชือกที่สั่ง เช่น ม้วน แกน กิโล เมตร');
             $table->integer('remaining_quantity')->comment('จำนวนที่เหลือยังไมไ่ด้ส่งผลิต ให้ start เท่า Order_quantity เมื่อเปิดใบสั่งขาย ให้มาลบกับอันนี้');
